+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+### Added
+-  `.getPaneName()` method
+### Fixed
+- Allow gl tiles to be added to a pane defined in options
+
 ## [0.0.12] - 2020-03-27
 
 ### Fixed